Converting doc encoding: iso8859-1 to utf-8.
\#fedora-devel suggested doing this in %setup instead of %build, but some
of the files are generated. So I'd either have to duplicate the code or
patch the generation process and specify utf-8 (something that I haven't
figured out how to do). For now this does eliminate the warnings from
rpmlint so I guess I'll leave it and see if it comes up in the review
process.